화면변경 할때 터지는 버그수정

+++ D3D/Client/pch.cpp
@@ -3,5 +3,5 @@
 
 unique_ptr<Core> core =make_unique<Core>();
 
-float WINDOW_WIDTH =1200.0f;
-float WINDOW_HEIGHT =800.0f;
+int WINDOW_WIDTH =1200;
+int WINDOW_HEIGHT =800;
